chore: 第三批 Ruff 历史质量债治理——修复 7 条规则 37 处违规（issue-25） (#53)
启用并修复 PLR0402(17) / UP041(4) / UP045(2) / UP036(1) / PIE790(2) /
RUF046(2) / RUF059(9)，共 37 处低中风险违规（13 个文件）。RUF059
逐处复核：均为解包后从未使用的变量，重命名为 _，不改断言语义。

- UP041: asyncio.TimeoutError -> TimeoutError（3.11 起内置别名）
- UP045: typing.Optional -> PEP604 X | None
- UP036: 删除 requires-python>=3.11 下恒真的版本判断死分支
- PIE790: 删除有 docstring 的类中多余 pass
- RUF046: 删除 round() 返回值上多余的 int()
- PLR0402: import x.y as y -> from x import y
- RUF059: 未使用解包变量重命名为 _

不改变公开行为；历史问题总数自 298 净下降至 261，新增 0。

+++ src/sts2_autotest/core/test_agent_runner.py
@@ -173,8 +173,8 @@ def _select_macos_window(
         ):
             continue
 
-        width = int(round(float(window.get("width", 0) or 0)))
-        height = int(round(float(window.get("height", 0) or 0)))
+        width = round(float(window.get("width", 0) or 0))
+        height = round(float(window.get("height", 0) or 0))
         window_id = int(window.get("window_id", 0) or 0)
         if window_id <= 0 or width <= 0 or height <= 0:
             continue
